requirements.txt
streamlit
pandas
gspread
oauth2client
python-barcode
Pillow